feat: unify notebook content under /learn
Every notebook with prose or dialogue now opens in a single content
reader at /learn/[id], regardless of whether it is a "book" or a
"story". The books/notebooks split is an implementation detail and no
longer leaks into the UI.

- Add has_content to NotebookSummary (proto + backend) — true when any
  scene has statements or conversations. /learn list cards route to the
  new reader when true.
- New /learn/[id] page: renders scenes with prose (statements) and/or
  dialogue (conversations), highlights defined expressions, and reuses
  the word-lookup popup for tap-to-define, save, and delete.
- Extract WordLookupPopup + useWordLookup hook into a shared component
  for reuse by the new reader.
- /notebooks/[id] "Read Book" button becomes "Read" and points to
  /learn/[id] whenever there is readable content.
- Delete /books and /books/[id] entirely (subsumed by /learn/[id]).
  Remove the Books card from the home page; update HomePage tests.
