+++ frontend/src/gen-protos/api/v1/quiz_pb.ts
@@ -11,7 +11,7 @@ import type { Message } from "@bufbuild/protobuf";
  * Describes the file api/v1/quiz.proto.
  */
 export const file_api_v1_quiz: GenFile = /*@__PURE__*/
-  fileDesc("ChFhcGkvdjEvcXVpei5wcm90bxIGYXBpLnYxIhcKFUdldFF1aXpPcHRpb25zUmVxdWVzdCJEChZHZXRRdWl6T3B0aW9uc1Jlc3BvbnNlEioKCW5vdGVib29rcxgBIAMoCzIXLmFwaS52MS5Ob3RlYm9va1N1bW1hcnkilgEKD05vdGVib29rU3VtbWFyeRITCgtub3RlYm9va19pZBgBIAEoCRIMCgRuYW1lGAIgASgJEhQKDHJldmlld19jb3VudBgDIAEoBRIMCgRraW5kGAQgASgJEhwKFHJldmVyc2VfcmV2aWV3X2NvdW50GAUgASgFEh4KFmV0eW1vbG9neV9yZXZpZXdfY291bnQYBiABKAUiTQoQU3RhcnRRdWl6UmVxdWVzdBIeCgxub3RlYm9va19pZHMYASADKAlCCLpIBZIBAggBEhkKEWluY2x1ZGVfdW5zdHVkaWVkGAIgASgIIjoKEVN0YXJ0UXVpelJlc3BvbnNlEiUKCmZsYXNoY2FyZHMYASADKAsyES5hcGkudjEuRmxhc2hjYXJkIk4KCUZsYXNoY2FyZBIPCgdub3RlX2lkGAEgASgDEg0KBWVudHJ5GAIgASgJEiEKCGV4YW1wbGVzGAMgAygLMg8uYXBpLnYxLkV4YW1wbGUiKAoHRXhhbXBsZRIMCgR0ZXh0GAEgASgJEg8KB3NwZWFrZXIYAiABKAkiqwEKCldvcmREZXRhaWwSDgoGb3JpZ2luGAEgASgJEhUKDXByb251bmNpYXRpb24YAiABKAkSFgoOcGFydF9vZl9zcGVlY2gYAyABKAkSEAoIc3lub255bXMYBCADKAkSEAoIYW50b255bXMYBSADKAkSDAoEbWVtbxgGIAEoCRIsCgxvcmlnaW5fcGFydHMYByADKAsyFi5hcGkudjEuV29yZE9yaWdpblBhcnQiUQoOV29yZE9yaWdpblBhcnQSDgoGb3JpZ2luGAEgASgJEgwKBHR5cGUYAiABKAkSEAoIbGFuZ3VhZ2UYAyABKAkSDwoHbWVhbmluZxgEIAEoCSJmChNTdWJtaXRBbnN3ZXJSZXF1ZXN0EhgKB25vdGVfaWQYASABKANCB7pIBCICIAASGwoGYW5zd2VyGAIgASgJQgu6SAhyBhABMgJcUxIYChByZXNwb25zZV90aW1lX21zGAMgASgDIq8BChRTdWJtaXRBbnN3ZXJSZXNwb25zZRIPCgdjb3JyZWN0GAEgASgIEg8KB21lYW5pbmcYAiABKAkSDgoGcmVhc29uGAMgASgJEicKC3dvcmRfZGV0YWlsGAQgASgLMhIuYXBpLnYxLldvcmREZXRhaWwSGAoQbmV4dF9yZXZpZXdfZGF0ZRgFIAEoCRISCgpsZWFybmVkX2F0GAYgASgJEg4KBmltYWdlcxgHIAMoCSJXChdTdGFydFJldmVyc2VRdWl6UmVxdWVzdBIeCgxub3RlYm9va19pZHMYASADKAlCCLpIBZIBAggBEhwKFGxpc3RfbWlzc2luZ19jb250ZXh0GAIgASgIIkgKGFN0YXJ0UmV2ZXJzZVF1aXpSZXNwb25zZRIsCgpmbGFzaGNhcmRzGAEgAygLMhguYXBpLnYxLlJldmVyc2VGbGFzaGNhcmQioAEKEFJldmVyc2VGbGFzaGNhcmQSDwoHbm90ZV9pZBgBIAEoAxIPCgdtZWFuaW5nGAIgASgJEikKCGNvbnRleHRzGAMgAygLMhcuYXBpLnYxLkNvbnRleHRTZW50ZW5jZRIVCg1ub3RlYm9va19uYW1lGAQgASgJEhMKC3N0b3J5X3RpdGxlGAUgASgJEhMKC3NjZW5lX3RpdGxlGAYgASgJIjoKD0NvbnRleHRTZW50ZW5jZRIPCgdjb250ZXh0GAEgASgJEhYKDm1hc2tlZF9jb250ZXh0GAIgASgJIm0KGlN1Ym1pdFJldmVyc2VBbnN3ZXJSZXF1ZXN0EhgKB25vdGVfaWQYASABKANCB7pIBCICIAASGwoGYW5zd2VyGAIgASgJQgu6SAhyBhABMgJcUxIYChByZXNwb25zZV90aW1lX21zGAMgASgDIvQBChtTdWJtaXRSZXZlcnNlQW5zd2VyUmVzcG9uc2USDwoHY29ycmVjdBgBIAEoCBISCgpleHByZXNzaW9uGAIgASgJEg8KB21lYW5pbmcYAyABKAkSDgoGcmVhc29uGAQgASgJEhAKCGNvbnRleHRzGAUgAygJEicKC3dvcmRfZGV0YWlsGAYgASgLMhIuYXBpLnYxLldvcmREZXRhaWwSFgoOY2xhc3NpZmljYXRpb24YByABKAkSGAoQbmV4dF9yZXZpZXdfZGF0ZRgIIAEoCRISCgpsZWFybmVkX2F0GAkgASgJEg4KBmltYWdlcxgKIAMoCSIaChhTdGFydEZyZWVmb3JtUXVpelJlcXVlc3Qi6wEKGVN0YXJ0RnJlZWZvcm1RdWl6UmVzcG9uc2USEgoKd29yZF9jb3VudBgBIAEoBRITCgtleHByZXNzaW9ucxgCIAMoCRJkChtleHByZXNzaW9uX25leHRfcmV2aWV3X2RhdGUYAyADKAsyPy5hcGkudjEuU3RhcnRGcmVlZm9ybVF1aXpSZXNwb25zZS5FeHByZXNzaW9uTmV4dFJldmlld0RhdGVFbnRyeRo/Ch1FeHByZXNzaW9uTmV4dFJldmlld0RhdGVFbnRyeRILCgNrZXkYASABKAkSDQoFdmFsdWUYAiABKAk6AjgBInAKG1N1Ym1pdEZyZWVmb3JtQW5zd2VyUmVxdWVzdBIZCgR3b3JkGAEgASgJQgu6SAhyBhABMgJcUxIcCgdtZWFuaW5nGAIgASgJQgu6SAhyBhABMgJcUxIYChByZXNwb25zZV90aW1lX21zGAMgASgDIv4BChxTdWJtaXRGcmVlZm9ybUFuc3dlclJlc3BvbnNlEg8KB2NvcnJlY3QYASABKAgSDAoEd29yZBgCIAEoCRIPCgdtZWFuaW5nGAMgASgJEg4KBnJlYXNvbhgEIAEoCRIPCgdjb250ZXh0GAUgASgJEhUKDW5vdGVib29rX25hbWUYBiABKAkSJwoLd29yZF9kZXRhaWwYByABKAsyEi5hcGkudjEuV29yZERldGFpbBIYChBuZXh0X3Jldmlld19kYXRlGAggASgJEhIKCmxlYXJuZWRfYXQYCSABKAkSDwoHbm90ZV9pZBgKIAEoAxIOCgZpbWFnZXMYCyADKAki0wEKFU92ZXJyaWRlQW5zd2VyUmVxdWVzdBIYCgdub3RlX2lkGAEgASgDQge6SAQiAiAAEiMKCXF1aXpfdHlwZRgCIAEoDjIQLmFwaS52MS5RdWl6VHlwZRIbCgpsZWFybmVkX2F0GAMgASgJQge6SARyAhAKEhkKDG1hcmtfY29ycmVjdBgEIAEoCEgAiAEBEh0KEG5leHRfcmV2aWV3X2RhdGUYBSABKAlIAYgBAUIPCg1fbWFya19jb3JyZWN0QhMKEV9uZXh0X3Jldmlld19kYXRlIosBChZPdmVycmlkZUFuc3dlclJlc3BvbnNlEhgKEG5leHRfcmV2aWV3X2RhdGUYASABKAkSGAoQb3JpZ2luYWxfcXVhbGl0eRgCIAEoBRIXCg9vcmlnaW5hbF9zdGF0dXMYAyABKAkSHgoWb3JpZ2luYWxfaW50ZXJ2YWxfZGF5cxgEIAEoBUoECAUQBiLQAQoZVW5kb092ZXJyaWRlQW5zd2VyUmVxdWVzdBIYCgdub3RlX2lkGAEgASgDQge6SAQiAiAAEiMKCXF1aXpfdHlwZRgCIAEoDjIQLmFwaS52MS5RdWl6VHlwZRIbCgpsZWFybmVkX2F0GAMgASgJQge6SARyAhAKEhgKEG9yaWdpbmFsX3F1YWxpdHkYBCABKAUSFwoPb3JpZ2luYWxfc3RhdHVzGAUgASgJEh4KFm9yaWdpbmFsX2ludGVydmFsX2RheXMYBiABKAVKBAgHEAgiRwoaVW5kb092ZXJyaWRlQW5zd2VyUmVzcG9uc2USDwoHY29ycmVjdBgBIAEoCBIYChBuZXh0X3Jldmlld19kYXRlGAIgASgJImQKD1NraXBXb3JkUmVxdWVzdBIYCgdub3RlX2lkGAEgASgDQge6SAQiAiAAEiMKCXF1aXpfdHlwZRgCIAEoDjIQLmFwaS52MS5RdWl6VHlwZRISCgpza2lwX3VudGlsGAMgASgJIhIKEFNraXBXb3JkUmVzcG9uc2UiUgoRUmVzdW1lV29yZFJlcXVlc3QSGAoHbm90ZV9pZBgBIAEoA0IHukgEIgIgABIjCglxdWl6X3R5cGUYAiABKA4yEC5hcGkudjEuUXVpelR5cGUiFAoSUmVzdW1lV29yZFJlc3BvbnNlInwKEUV0eW1vbG9neVF1aXpDYXJkEg8KB2NhcmRfaWQYASABKAMSDgoGb3JpZ2luGAIgASgJEgwKBHR5cGUYAyABKAkSEAoIbGFuZ3VhZ2UYBCABKAkSDwoHbWVhbmluZxgFIAEoCRIVCg1ub3RlYm9va19uYW1lGAYgASgJIn8KGVN0YXJ0RXR5bW9sb2d5UXVpelJlcXVlc3QSHgoWZXR5bW9sb2d5X25vdGVib29rX2lkcxgBIAMoCRInCgRtb2RlGAMgASgOMhkuYXBpLnYxLkV0eW1vbG9neVF1aXpNb2RlEhkKEWluY2x1ZGVfdW5zdHVkaWVkGAQgASgIIkYKGlN0YXJ0RXR5bW9sb2d5UXVpelJlc3BvbnNlEigKBWNhcmRzGAEgAygLMhkuYXBpLnYxLkV0eW1vbG9neVF1aXpDYXJkIncKJFN1Ym1pdEV0eW1vbG9neVN0YW5kYXJkQW5zd2VyUmVxdWVzdBIYCgdjYXJkX2lkGAEgASgDQge6SAQiAiAAEhsKBmFuc3dlchgCIAEoCUILukgIcgYQATICXFMSGAoQcmVzcG9uc2VfdGltZV9tcxgDIAEoAyKgAQolU3VibWl0RXR5bW9sb2d5U3RhbmRhcmRBbnN3ZXJSZXNwb25zZRIPCgdjb3JyZWN0GAEgASgIEg4KBnJlYXNvbhgCIAEoCRIXCg9jb3JyZWN0X21lYW5pbmcYAyABKAkSGAoQbmV4dF9yZXZpZXdfZGF0ZRgEIAEoCRISCgpsZWFybmVkX2F0GAUgASgJEg8KB25vdGVfaWQYBiABKAMidgojU3VibWl0RXR5bW9sb2d5UmV2ZXJzZUFuc3dlclJlcXVlc3QSGAoHY2FyZF9pZBgBIAEoA0IHukgEIgIgABIbCgZhbnN3ZXIYAiABKAlCC7pICHIGEAEyAlxTEhgKEHJlc3BvbnNlX3RpbWVfbXMYAyABKAMivgEKJFN1Ym1pdEV0eW1vbG9neVJldmVyc2VBbnN3ZXJSZXNwb25zZRIPCgdjb3JyZWN0GAEgASgIEg4KBnJlYXNvbhgCIAEoCRIWCg5jb3JyZWN0X29yaWdpbhgDIAEoCRIMCgR0eXBlGAQgASgJEhAKCGxhbmd1YWdlGAUgASgJEhgKEG5leHRfcmV2aWV3X2RhdGUYBiABKAkSEgoKbGVhcm5lZF9hdBgHIAEoCRIPCgdub3RlX2lkGAggASgDIkMKIVN0YXJ0RXR5bW9sb2d5RnJlZWZvcm1RdWl6UmVxdWVzdBIeChZldHltb2xvZ3lfbm90ZWJvb2tfaWRzGAEgAygJIskBCiJTdGFydEV0eW1vbG9neUZyZWVmb3JtUXVpelJlc3BvbnNlEg8KB29yaWdpbnMYASADKAkSWgoRbmV4dF9yZXZpZXdfZGF0ZXMYAiADKAsyPy5hcGkudjEuU3RhcnRFdHltb2xvZ3lGcmVlZm9ybVF1aXpSZXNwb25zZS5OZXh0UmV2aWV3RGF0ZXNFbnRyeRo2ChROZXh0UmV2aWV3RGF0ZXNFbnRyeRILCgNrZXkYASABKAkSDQoFdmFsdWUYAiABKAk6AjgBInsKJFN1Ym1pdEV0eW1vbG9neUZyZWVmb3JtQW5zd2VyUmVxdWVzdBIbCgZvcmlnaW4YASABKAlCC7pICHIGEAEyAlxTEhwKB21lYW5pbmcYAiABKAlCC7pICHIGEAEyAlxTEhgKEHJlc3BvbnNlX3RpbWVfbXMYAyABKAMi1wEKJVN1Ym1pdEV0eW1vbG9neUZyZWVmb3JtQW5zd2VyUmVzcG9uc2USDwoHY29ycmVjdBgBIAEoCBIOCgZyZWFzb24YAiABKAkSFwoPY29ycmVjdF9tZWFuaW5nGAMgASgJEgwKBHR5cGUYBCABKAkSEAoIbGFuZ3VhZ2UYBSABKAkSFQoNbm90ZWJvb2tfbmFtZRgGIAEoCRIYChBuZXh0X3Jldmlld19kYXRlGAcgASgJEhIKCmxlYXJuZWRfYXQYCCABKAkSDwoHbm90ZV9pZBgJIAEoAyrRAQoIUXVpelR5cGUSGQoVUVVJWl9UWVBFX1VOU1BFQ0lGSUVEEAASFgoSUVVJWl9UWVBFX1NUQU5EQVJEEAESFQoRUVVJWl9UWVBFX1JFVkVSU0UQAhIWChJRVUlaX1RZUEVfRlJFRUZPUk0QAxIgChxRVUlaX1RZUEVfRVRZTU9MT0dZX1NUQU5EQVJEEAQSHwobUVVJWl9UWVBFX0VUWU1PTE9HWV9SRVZFUlNFEAUSIAocUVVJWl9UWVBFX0VUWU1PTE9HWV9GUkVFRk9STRAGKnsKEUV0eW1vbG9neVF1aXpNb2RlEiMKH0VUWU1PTE9HWV9RVUlaX01PREVfVU5TUEVDSUZJRUQQABIgChxFVFlNT0xPR1lfUVVJWl9NT0RFX1NUQU5EQVJEEAESHwobRVRZTU9MT0dZX1FVSVpfTU9ERV9SRVZFUlNFEAIy2gsKC1F1aXpTZXJ2aWNlEk8KDkdldFF1aXpPcHRpb25zEh0uYXBpLnYxLkdldFF1aXpPcHRpb25zUmVxdWVzdBoeLmFwaS52MS5HZXRRdWl6T3B0aW9uc1Jlc3BvbnNlEkAKCVN0YXJ0UXVpehIYLmFwaS52MS5TdGFydFF1aXpSZXF1ZXN0GhkuYXBpLnYxLlN0YXJ0UXVpelJlc3BvbnNlEkkKDFN1Ym1pdEFuc3dlchIbLmFwaS52MS5TdWJtaXRBbnN3ZXJSZXF1ZXN0GhwuYXBpLnYxLlN1Ym1pdEFuc3dlclJlc3BvbnNlElUKEFN0YXJ0UmV2ZXJzZVF1aXoSHy5hcGkudjEuU3RhcnRSZXZlcnNlUXVpelJlcXVlc3QaIC5hcGkudjEuU3RhcnRSZXZlcnNlUXVpelJlc3BvbnNlEl4KE1N1Ym1pdFJldmVyc2VBbnN3ZXISIi5hcGkudjEuU3VibWl0UmV2ZXJzZUFuc3dlclJlcXVlc3QaIy5hcGkudjEuU3VibWl0UmV2ZXJzZUFuc3dlclJlc3BvbnNlElgKEVN0YXJ0RnJlZWZvcm1RdWl6EiAuYXBpLnYxLlN0YXJ0RnJlZWZvcm1RdWl6UmVxdWVzdBohLmFwaS52MS5TdGFydEZyZWVmb3JtUXVpelJlc3BvbnNlEmEKFFN1Ym1pdEZyZWVmb3JtQW5zd2VyEiMuYXBpLnYxLlN1Ym1pdEZyZWVmb3JtQW5zd2VyUmVxdWVzdBokLmFwaS52MS5TdWJtaXRGcmVlZm9ybUFuc3dlclJlc3BvbnNlEk8KDk92ZXJyaWRlQW5zd2VyEh0uYXBpLnYxLk92ZXJyaWRlQW5zd2VyUmVxdWVzdBoeLmFwaS52MS5PdmVycmlkZUFuc3dlclJlc3BvbnNlElsKElVuZG9PdmVycmlkZUFuc3dlchIhLmFwaS52MS5VbmRvT3ZlcnJpZGVBbnN3ZXJSZXF1ZXN0GiIuYXBpLnYxLlVuZG9PdmVycmlkZUFuc3dlclJlc3BvbnNlEj0KCFNraXBXb3JkEhcuYXBpLnYxLlNraXBXb3JkUmVxdWVzdBoYLmFwaS52MS5Ta2lwV29yZFJlc3BvbnNlEkMKClJlc3VtZVdvcmQSGS5hcGkudjEuUmVzdW1lV29yZFJlcXVlc3QaGi5hcGkudjEuUmVzdW1lV29yZFJlc3BvbnNlElsKElN0YXJ0RXR5bW9sb2d5UXVpehIhLmFwaS52MS5TdGFydEV0eW1vbG9neVF1aXpSZXF1ZXN0GiIuYXBpLnYxLlN0YXJ0RXR5bW9sb2d5UXVpelJlc3BvbnNlEnwKHVN1Ym1pdEV0eW1vbG9neVN0YW5kYXJkQW5zd2VyEiwuYXBpLnYxLlN1Ym1pdEV0eW1vbG9neVN0YW5kYXJkQW5zd2VyUmVxdWVzdBotLmFwaS52MS5TdWJtaXRFdHltb2xvZ3lTdGFuZGFyZEFuc3dlclJlc3BvbnNlEnkKHFN1Ym1pdEV0eW1vbG9neVJldmVyc2VBbnN3ZXISKy5hcGkudjEuU3VibWl0RXR5bW9sb2d5UmV2ZXJzZUFuc3dlclJlcXVlc3QaLC5hcGkudjEuU3VibWl0RXR5bW9sb2d5UmV2ZXJzZUFuc3dlclJlc3BvbnNlEnMKGlN0YXJ0RXR5bW9sb2d5RnJlZWZvcm1RdWl6EikuYXBpLnYxLlN0YXJ0RXR5bW9sb2d5RnJlZWZvcm1RdWl6UmVxdWVzdBoqLmFwaS52MS5TdGFydEV0eW1vbG9neUZyZWVmb3JtUXVpelJlc3BvbnNlEnwKHVN1Ym1pdEV0eW1vbG9neUZyZWVmb3JtQW5zd2VyEiwuYXBpLnYxLlN1Ym1pdEV0eW1vbG9neUZyZWVmb3JtQW5zd2VyUmVxdWVzdBotLmFwaS52MS5TdWJtaXRFdHltb2xvZ3lGcmVlZm9ybUFuc3dlclJlc3BvbnNlQjhaNmdpdGh1Yi5jb20vYXQtaXNoaWthd2EvbGFuZ25lci9nZW4tcHJvdG9zL2FwaS92MTthcGl2MWIGcHJvdG8z", [file_buf_validate_validate]);
+  fileDesc("ChFhcGkvdjEvcXVpei5wcm90bxIGYXBpLnYxIhcKFUdldFF1aXpPcHRpb25zUmVxdWVzdCJEChZHZXRRdWl6T3B0aW9uc1Jlc3BvbnNlEioKCW5vdGVib29rcxgBIAMoCzIXLmFwaS52MS5Ob3RlYm9va1N1bW1hcnkiqwEKD05vdGVib29rU3VtbWFyeRITCgtub3RlYm9va19pZBgBIAEoCRIMCgRuYW1lGAIgASgJEhQKDHJldmlld19jb3VudBgDIAEoBRIMCgRraW5kGAQgASgJEhwKFHJldmVyc2VfcmV2aWV3X2NvdW50GAUgASgFEh4KFmV0eW1vbG9neV9yZXZpZXdfY291bnQYBiABKAUSEwoLaGFzX2NvbnRlbnQYByABKAgiTQoQU3RhcnRRdWl6UmVxdWVzdBIeCgxub3RlYm9va19pZHMYASADKAlCCLpIBZIBAggBEhkKEWluY2x1ZGVfdW5zdHVkaWVkGAIgASgIIjoKEVN0YXJ0UXVpelJlc3BvbnNlEiUKCmZsYXNoY2FyZHMYASADKAsyES5hcGkudjEuRmxhc2hjYXJkIk4KCUZsYXNoY2FyZBIPCgdub3RlX2lkGAEgASgDEg0KBWVudHJ5GAIgASgJEiEKCGV4YW1wbGVzGAMgAygLMg8uYXBpLnYxLkV4YW1wbGUiKAoHRXhhbXBsZRIMCgR0ZXh0GAEgASgJEg8KB3NwZWFrZXIYAiABKAkiqwEKCldvcmREZXRhaWwSDgoGb3JpZ2luGAEgASgJEhUKDXByb251bmNpYXRpb24YAiABKAkSFgoOcGFydF9vZl9zcGVlY2gYAyABKAkSEAoIc3lub255bXMYBCADKAkSEAoIYW50b255bXMYBSADKAkSDAoEbWVtbxgGIAEoCRIsCgxvcmlnaW5fcGFydHMYByADKAsyFi5hcGkudjEuV29yZE9yaWdpblBhcnQiUQoOV29yZE9yaWdpblBhcnQSDgoGb3JpZ2luGAEgASgJEgwKBHR5cGUYAiABKAkSEAoIbGFuZ3VhZ2UYAyABKAkSDwoHbWVhbmluZxgEIAEoCSJmChNTdWJtaXRBbnN3ZXJSZXF1ZXN0EhgKB25vdGVfaWQYASABKANCB7pIBCICIAASGwoGYW5zd2VyGAIgASgJQgu6SAhyBhABMgJcUxIYChByZXNwb25zZV90aW1lX21zGAMgASgDIq8BChRTdWJtaXRBbnN3ZXJSZXNwb25zZRIPCgdjb3JyZWN0GAEgASgIEg8KB21lYW5pbmcYAiABKAkSDgoGcmVhc29uGAMgASgJEicKC3dvcmRfZGV0YWlsGAQgASgLMhIuYXBpLnYxLldvcmREZXRhaWwSGAoQbmV4dF9yZXZpZXdfZGF0ZRgFIAEoCRISCgpsZWFybmVkX2F0GAYgASgJEg4KBmltYWdlcxgHIAMoCSJXChdTdGFydFJldmVyc2VRdWl6UmVxdWVzdBIeCgxub3RlYm9va19pZHMYASADKAlCCLpIBZIBAggBEhwKFGxpc3RfbWlzc2luZ19jb250ZXh0GAIgASgIIkgKGFN0YXJ0UmV2ZXJzZVF1aXpSZXNwb25zZRIsCgpmbGFzaGNhcmRzGAEgAygLMhguYXBpLnYxLlJldmVyc2VGbGFzaGNhcmQioAEKEFJldmVyc2VGbGFzaGNhcmQSDwoHbm90ZV9pZBgBIAEoAxIPCgdtZWFuaW5nGAIgASgJEikKCGNvbnRleHRzGAMgAygLMhcuYXBpLnYxLkNvbnRleHRTZW50ZW5jZRIVCg1ub3RlYm9va19uYW1lGAQgASgJEhMKC3N0b3J5X3RpdGxlGAUgASgJEhMKC3NjZW5lX3RpdGxlGAYgASgJIjoKD0NvbnRleHRTZW50ZW5jZRIPCgdjb250ZXh0GAEgASgJEhYKDm1hc2tlZF9jb250ZXh0GAIgASgJIm0KGlN1Ym1pdFJldmVyc2VBbnN3ZXJSZXF1ZXN0EhgKB25vdGVfaWQYASABKANCB7pIBCICIAASGwoGYW5zd2VyGAIgASgJQgu6SAhyBhABMgJcUxIYChByZXNwb25zZV90aW1lX21zGAMgASgDIvQBChtTdWJtaXRSZXZlcnNlQW5zd2VyUmVzcG9uc2USDwoHY29ycmVjdBgBIAEoCBISCgpleHByZXNzaW9uGAIgASgJEg8KB21lYW5pbmcYAyABKAkSDgoGcmVhc29uGAQgASgJEhAKCGNvbnRleHRzGAUgAygJEicKC3dvcmRfZGV0YWlsGAYgASgLMhIuYXBpLnYxLldvcmREZXRhaWwSFgoOY2xhc3NpZmljYXRpb24YByABKAkSGAoQbmV4dF9yZXZpZXdfZGF0ZRgIIAEoCRISCgpsZWFybmVkX2F0GAkgASgJEg4KBmltYWdlcxgKIAMoCSIaChhTdGFydEZyZWVmb3JtUXVpelJlcXVlc3Qi6wEKGVN0YXJ0RnJlZWZvcm1RdWl6UmVzcG9uc2USEgoKd29yZF9jb3VudBgBIAEoBRITCgtleHByZXNzaW9ucxgCIAMoCRJkChtleHByZXNzaW9uX25leHRfcmV2aWV3X2RhdGUYAyADKAsyPy5hcGkudjEuU3RhcnRGcmVlZm9ybVF1aXpSZXNwb25zZS5FeHByZXNzaW9uTmV4dFJldmlld0RhdGVFbnRyeRo/Ch1FeHByZXNzaW9uTmV4dFJldmlld0RhdGVFbnRyeRILCgNrZXkYASABKAkSDQoFdmFsdWUYAiABKAk6AjgBInAKG1N1Ym1pdEZyZWVmb3JtQW5zd2VyUmVxdWVzdBIZCgR3b3JkGAEgASgJQgu6SAhyBhABMgJcUxIcCgdtZWFuaW5nGAIgASgJQgu6SAhyBhABMgJcUxIYChByZXNwb25zZV90aW1lX21zGAMgASgDIv4BChxTdWJtaXRGcmVlZm9ybUFuc3dlclJlc3BvbnNlEg8KB2NvcnJlY3QYASABKAgSDAoEd29yZBgCIAEoCRIPCgdtZWFuaW5nGAMgASgJEg4KBnJlYXNvbhgEIAEoCRIPCgdjb250ZXh0GAUgASgJEhUKDW5vdGVib29rX25hbWUYBiABKAkSJwoLd29yZF9kZXRhaWwYByABKAsyEi5hcGkudjEuV29yZERldGFpbBIYChBuZXh0X3Jldmlld19kYXRlGAggASgJEhIKCmxlYXJuZWRfYXQYCSABKAkSDwoHbm90ZV9pZBgKIAEoAxIOCgZpbWFnZXMYCyADKAki0wEKFU92ZXJyaWRlQW5zd2VyUmVxdWVzdBIYCgdub3RlX2lkGAEgASgDQge6SAQiAiAAEiMKCXF1aXpfdHlwZRgCIAEoDjIQLmFwaS52MS5RdWl6VHlwZRIbCgpsZWFybmVkX2F0GAMgASgJQge6SARyAhAKEhkKDG1hcmtfY29ycmVjdBgEIAEoCEgAiAEBEh0KEG5leHRfcmV2aWV3X2RhdGUYBSABKAlIAYgBAUIPCg1fbWFya19jb3JyZWN0QhMKEV9uZXh0X3Jldmlld19kYXRlIosBChZPdmVycmlkZUFuc3dlclJlc3BvbnNlEhgKEG5leHRfcmV2aWV3X2RhdGUYASABKAkSGAoQb3JpZ2luYWxfcXVhbGl0eRgCIAEoBRIXCg9vcmlnaW5hbF9zdGF0dXMYAyABKAkSHgoWb3JpZ2luYWxfaW50ZXJ2YWxfZGF5cxgEIAEoBUoECAUQBiLQAQoZVW5kb092ZXJyaWRlQW5zd2VyUmVxdWVzdBIYCgdub3RlX2lkGAEgASgDQge6SAQiAiAAEiMKCXF1aXpfdHlwZRgCIAEoDjIQLmFwaS52MS5RdWl6VHlwZRIbCgpsZWFybmVkX2F0GAMgASgJQge6SARyAhAKEhgKEG9yaWdpbmFsX3F1YWxpdHkYBCABKAUSFwoPb3JpZ2luYWxfc3RhdHVzGAUgASgJEh4KFm9yaWdpbmFsX2ludGVydmFsX2RheXMYBiABKAVKBAgHEAgiRwoaVW5kb092ZXJyaWRlQW5zd2VyUmVzcG9uc2USDwoHY29ycmVjdBgBIAEoCBIYChBuZXh0X3Jldmlld19kYXRlGAIgASgJImQKD1NraXBXb3JkUmVxdWVzdBIYCgdub3RlX2lkGAEgASgDQge6SAQiAiAAEiMKCXF1aXpfdHlwZRgCIAEoDjIQLmFwaS52MS5RdWl6VHlwZRISCgpza2lwX3VudGlsGAMgASgJIhIKEFNraXBXb3JkUmVzcG9uc2UiUgoRUmVzdW1lV29yZFJlcXVlc3QSGAoHbm90ZV9pZBgBIAEoA0IHukgEIgIgABIjCglxdWl6X3R5cGUYAiABKA4yEC5hcGkudjEuUXVpelR5cGUiFAoSUmVzdW1lV29yZFJlc3BvbnNlInwKEUV0eW1vbG9neVF1aXpDYXJkEg8KB2NhcmRfaWQYASABKAMSDgoGb3JpZ2luGAIgASgJEgwKBHR5cGUYAyABKAkSEAoIbGFuZ3VhZ2UYBCABKAkSDwoHbWVhbmluZxgFIAEoCRIVCg1ub3RlYm9va19uYW1lGAYgASgJIn8KGVN0YXJ0RXR5bW9sb2d5UXVpelJlcXVlc3QSHgoWZXR5bW9sb2d5X25vdGVib29rX2lkcxgBIAMoCRInCgRtb2RlGAMgASgOMhkuYXBpLnYxLkV0eW1vbG9neVF1aXpNb2RlEhkKEWluY2x1ZGVfdW5zdHVkaWVkGAQgASgIIkYKGlN0YXJ0RXR5bW9sb2d5UXVpelJlc3BvbnNlEigKBWNhcmRzGAEgAygLMhkuYXBpLnYxLkV0eW1vbG9neVF1aXpDYXJkIncKJFN1Ym1pdEV0eW1vbG9neVN0YW5kYXJkQW5zd2VyUmVxdWVzdBIYCgdjYXJkX2lkGAEgASgDQge6SAQiAiAAEhsKBmFuc3dlchgCIAEoCUILukgIcgYQATICXFMSGAoQcmVzcG9uc2VfdGltZV9tcxgDIAEoAyKgAQolU3VibWl0RXR5bW9sb2d5U3RhbmRhcmRBbnN3ZXJSZXNwb25zZRIPCgdjb3JyZWN0GAEgASgIEg4KBnJlYXNvbhgCIAEoCRIXCg9jb3JyZWN0X21lYW5pbmcYAyABKAkSGAoQbmV4dF9yZXZpZXdfZGF0ZRgEIAEoCRISCgpsZWFybmVkX2F0GAUgASgJEg8KB25vdGVfaWQYBiABKAMidgojU3VibWl0RXR5bW9sb2d5UmV2ZXJzZUFuc3dlclJlcXVlc3QSGAoHY2FyZF9pZBgBIAEoA0IHukgEIgIgABIbCgZhbnN3ZXIYAiABKAlCC7pICHIGEAEyAlxTEhgKEHJlc3BvbnNlX3RpbWVfbXMYAyABKAMivgEKJFN1Ym1pdEV0eW1vbG9neVJldmVyc2VBbnN3ZXJSZXNwb25zZRIPCgdjb3JyZWN0GAEgASgIEg4KBnJlYXNvbhgCIAEoCRIWCg5jb3JyZWN0X29yaWdpbhgDIAEoCRIMCgR0eXBlGAQgASgJEhAKCGxhbmd1YWdlGAUgASgJEhgKEG5leHRfcmV2aWV3X2RhdGUYBiABKAkSEgoKbGVhcm5lZF9hdBgHIAEoCRIPCgdub3RlX2lkGAggASgDIkMKIVN0YXJ0RXR5bW9sb2d5RnJlZWZvcm1RdWl6UmVxdWVzdBIeChZldHltb2xvZ3lfbm90ZWJvb2tfaWRzGAEgAygJIskBCiJTdGFydEV0eW1vbG9neUZyZWVmb3JtUXVpelJlc3BvbnNlEg8KB29yaWdpbnMYASADKAkSWgoRbmV4dF9yZXZpZXdfZGF0ZXMYAiADKAsyPy5hcGkudjEuU3RhcnRFdHltb2xvZ3lGcmVlZm9ybVF1aXpSZXNwb25zZS5OZXh0UmV2aWV3RGF0ZXNFbnRyeRo2ChROZXh0UmV2aWV3RGF0ZXNFbnRyeRILCgNrZXkYASABKAkSDQoFdmFsdWUYAiABKAk6AjgBInsKJFN1Ym1pdEV0eW1vbG9neUZyZWVmb3JtQW5zd2VyUmVxdWVzdBIbCgZvcmlnaW4YASABKAlCC7pICHIGEAEyAlxTEhwKB21lYW5pbmcYAiABKAlCC7pICHIGEAEyAlxTEhgKEHJlc3BvbnNlX3RpbWVfbXMYAyABKAMi1wEKJVN1Ym1pdEV0eW1vbG9neUZyZWVmb3JtQW5zd2VyUmVzcG9uc2USDwoHY29ycmVjdBgBIAEoCBIOCgZyZWFzb24YAiABKAkSFwoPY29ycmVjdF9tZWFuaW5nGAMgASgJEgwKBHR5cGUYBCABKAkSEAoIbGFuZ3VhZ2UYBSABKAkSFQoNbm90ZWJvb2tfbmFtZRgGIAEoCRIYChBuZXh0X3Jldmlld19kYXRlGAcgASgJEhIKCmxlYXJuZWRfYXQYCCABKAkSDwoHbm90ZV9pZBgJIAEoAyrRAQoIUXVpelR5cGUSGQoVUVVJWl9UWVBFX1VOU1BFQ0lGSUVEEAASFgoSUVVJWl9UWVBFX1NUQU5EQVJEEAESFQoRUVVJWl9UWVBFX1JFVkVSU0UQAhIWChJRVUlaX1RZUEVfRlJFRUZPUk0QAxIgChxRVUlaX1RZUEVfRVRZTU9MT0dZX1NUQU5EQVJEEAQSHwobUVVJWl9UWVBFX0VUWU1PTE9HWV9SRVZFUlNFEAUSIAocUVVJWl9UWVBFX0VUWU1PTE9HWV9GUkVFRk9STRAGKnsKEUV0eW1vbG9neVF1aXpNb2RlEiMKH0VUWU1PTE9HWV9RVUlaX01PREVfVU5TUEVDSUZJRUQQABIgChxFVFlNT0xPR1lfUVVJWl9NT0RFX1NUQU5EQVJEEAESHwobRVRZTU9MT0dZX1FVSVpfTU9ERV9SRVZFUlNFEAIy2gsKC1F1aXpTZXJ2aWNlEk8KDkdldFF1aXpPcHRpb25zEh0uYXBpLnYxLkdldFF1aXpPcHRpb25zUmVxdWVzdBoeLmFwaS52MS5HZXRRdWl6T3B0aW9uc1Jlc3BvbnNlEkAKCVN0YXJ0UXVpehIYLmFwaS52MS5TdGFydFF1aXpSZXF1ZXN0GhkuYXBpLnYxLlN0YXJ0UXVpelJlc3BvbnNlEkkKDFN1Ym1pdEFuc3dlchIbLmFwaS52MS5TdWJtaXRBbnN3ZXJSZXF1ZXN0GhwuYXBpLnYxLlN1Ym1pdEFuc3dlclJlc3BvbnNlElUKEFN0YXJ0UmV2ZXJzZVF1aXoSHy5hcGkudjEuU3RhcnRSZXZlcnNlUXVpelJlcXVlc3QaIC5hcGkudjEuU3RhcnRSZXZlcnNlUXVpelJlc3BvbnNlEl4KE1N1Ym1pdFJldmVyc2VBbnN3ZXISIi5hcGkudjEuU3VibWl0UmV2ZXJzZUFuc3dlclJlcXVlc3QaIy5hcGkudjEuU3VibWl0UmV2ZXJzZUFuc3dlclJlc3BvbnNlElgKEVN0YXJ0RnJlZWZvcm1RdWl6EiAuYXBpLnYxLlN0YXJ0RnJlZWZvcm1RdWl6UmVxdWVzdBohLmFwaS52MS5TdGFydEZyZWVmb3JtUXVpelJlc3BvbnNlEmEKFFN1Ym1pdEZyZWVmb3JtQW5zd2VyEiMuYXBpLnYxLlN1Ym1pdEZyZWVmb3JtQW5zd2VyUmVxdWVzdBokLmFwaS52MS5TdWJtaXRGcmVlZm9ybUFuc3dlclJlc3BvbnNlEk8KDk92ZXJyaWRlQW5zd2VyEh0uYXBpLnYxLk92ZXJyaWRlQW5zd2VyUmVxdWVzdBoeLmFwaS52MS5PdmVycmlkZUFuc3dlclJlc3BvbnNlElsKElVuZG9PdmVycmlkZUFuc3dlchIhLmFwaS52MS5VbmRvT3ZlcnJpZGVBbnN3ZXJSZXF1ZXN0GiIuYXBpLnYxLlVuZG9PdmVycmlkZUFuc3dlclJlc3BvbnNlEj0KCFNraXBXb3JkEhcuYXBpLnYxLlNraXBXb3JkUmVxdWVzdBoYLmFwaS52MS5Ta2lwV29yZFJlc3BvbnNlEkMKClJlc3VtZVdvcmQSGS5hcGkudjEuUmVzdW1lV29yZFJlcXVlc3QaGi5hcGkudjEuUmVzdW1lV29yZFJlc3BvbnNlElsKElN0YXJ0RXR5bW9sb2d5UXVpehIhLmFwaS52MS5TdGFydEV0eW1vbG9neVF1aXpSZXF1ZXN0GiIuYXBpLnYxLlN0YXJ0RXR5bW9sb2d5UXVpelJlc3BvbnNlEnwKHVN1Ym1pdEV0eW1vbG9neVN0YW5kYXJkQW5zd2VyEiwuYXBpLnYxLlN1Ym1pdEV0eW1vbG9neVN0YW5kYXJkQW5zd2VyUmVxdWVzdBotLmFwaS52MS5TdWJtaXRFdHltb2xvZ3lTdGFuZGFyZEFuc3dlclJlc3BvbnNlEnkKHFN1Ym1pdEV0eW1vbG9neVJldmVyc2VBbnN3ZXISKy5hcGkudjEuU3VibWl0RXR5bW9sb2d5UmV2ZXJzZUFuc3dlclJlcXVlc3QaLC5hcGkudjEuU3VibWl0RXR5bW9sb2d5UmV2ZXJzZUFuc3dlclJlc3BvbnNlEnMKGlN0YXJ0RXR5bW9sb2d5RnJlZWZvcm1RdWl6EikuYXBpLnYxLlN0YXJ0RXR5bW9sb2d5RnJlZWZvcm1RdWl6UmVxdWVzdBoqLmFwaS52MS5TdGFydEV0eW1vbG9neUZyZWVmb3JtUXVpelJlc3BvbnNlEnwKHVN1Ym1pdEV0eW1vbG9neUZyZWVmb3JtQW5zd2VyEiwuYXBpLnYxLlN1Ym1pdEV0eW1vbG9neUZyZWVmb3JtQW5zd2VyUmVxdWVzdBotLmFwaS52MS5TdWJtaXRFdHltb2xvZ3lGcmVlZm9ybUFuc3dlclJlc3BvbnNlQjhaNmdpdGh1Yi5jb20vYXQtaXNoaWthd2EvbGFuZ25lci9nZW4tcHJvdG9zL2FwaS92MTthcGl2MWIGcHJvdG8z", [file_buf_validate_validate]);
 
 /**
  * @generated from message api.v1.GetQuizOptionsRequest
@@ -76,6 +76,14 @@ export type NotebookSummary = Message<"api.v1.NotebookSummary"> & {
    * @generated from field: int32 etymology_review_count = 6;
    */
   etymologyReviewCount: number;
+
+  /**
+   * has_content is true when any scene in the notebook has statements or
+   * conversations — i.e., there is prose/dialogue to read, not just flashcards.
+   *
+   * @generated from field: bool has_content = 7;
+   */
+  hasContent: boolean;
 };
 
 /**
